+++ examples/default/eventcatalog.config.js
@@ -28,5 +28,19 @@ export default {
   },
   llmsTxt: {
     enabled: true,
-  }
+  },
+  chat: {
+    enabled: true,
+    similarityResults: 50,
+    max_tokens: 4096,
+    // 'Llama-3.2-3B-Instruct-q4f16_1-MLC is also good
+    model: 'Hermes-3-Llama-3.2-3B-q4f16_1-MLC'
+  },
+  generators: [
+    [
+      "@eventcatalog/generator-ai", {
+        splitMarkdownFiles: false,
+      }
+    ],
+  ],
 };